Extract user schema creation to separate file
